Split up the files.
