+++ types.go
@@ -1,28 +1,30 @@
 package subscribe
 
 import (
+	"errors"
 	"sync"
 	"time"
 )
 
-// Error enables constant errors.
-type Error string
-
-// Error allows a string to satisfy the error type.
-func (e Error) Error() string {
-	return string(e)
-}
-
-const (
+var (
 	// ErrorSubscriberNotFound is returned any time a requested subscriber does not exist.
-	ErrorSubscriberNotFound = Error("subscriber not found")
+	ErrorSubscriberNotFound = errors.New("subscriber not found")
 	// ErrorEventNotFound is returned when a requested event has not been created.
-	ErrorEventNotFound = Error("event not found")
+	ErrorEventNotFound = errors.New("event not found")
 	// ErrorEventExists is returned when a new event with an existing name is created.
-	ErrorEventExists = Error("event already exists")
+	ErrorEventExists = errors.New("event already exists")
 )
 
-type subscriberEvents map[string]time.Time
+// SubEventInfo contains the pause time and rules for a subscriber's event subscription.
+// Rules are unused by the library and available for consumers.
+type SubEventInfo struct {
+	Pause time.Time
+	Rules []string
+}
+
+// SubEvents represent's a subscriber's list of event subscriptions.
+// Each subscription has a unique name and some meta data attached.
+type SubEvents map[string]SubEventInfo
 
 // Subscriber describes the contact info and subscriptions for a person.
 type Subscriber struct {
@@ -31,16 +33,22 @@ type Subscriber struct {
 	// Contact is the contact info used in the API to send the subscriber a notification.
 	Contact string `json:"contact"`
 	// Events is a list of events the subscriber is subscribed to, including a cooldown/pause time.
-	Events subscriberEvents `json:"events"`
+	Events SubEvents `json:"events"`
 	// This is just extra data that can be used to make the user special.
 	Admin bool `json:"is_admin"`
 	// Ignored will exclude a user from GetSubscribers().
 	Ignored bool `json:"ignored"`
-	// sync.RWMutex Locks/UnlocksE vents map
+	// sync.RWMutex Locks/Unlocks Events map
 	sync.RWMutex
 }
 
-type subscribeEvents map[string]map[string]string
+// Rules is arbitrary data that can be stored with an event.
+type Rules map[string]string
+
+// Events represents the map of tracked global events.
+// This is an arbitrary list that can be used to filter
+// notifications in a consuming application.
+type Events map[string]Rules
 
 // Subscribe is the data needed to initialize this module.
 type Subscribe struct {
@@ -50,8 +58,8 @@ type Subscribe struct {
 	// stateFile is the db location, like: /usr/local/var/lib/motifini/subscribers.json
 	stateFile string
 	// Events stores a list of arbitrary events. Use the included methods to interact with it.
-	// This does not affect GetSubscribers().
-	Events subscribeEvents `json:"events"`
+	// This does not affect GetSubscribers(). Use the data here as a filter in your app.
+	Events Events `json:"events"`
 	// Subscribers is a list of all Subscribers.
 	Subscribers []*Subscriber `json:"subscribers"`
 	// sync.RWMutex locks and unlocks the Events map
